use usize for integers

+++ src/scanner/token.rs
@@ -13,7 +13,7 @@ pub struct Token {
     pub token_type: TokenType,
     pub lexeme: String,
     pub literal: Option<Literal>,
-    pub line: u64,
+    pub line: usize,
 }
 
 impl fmt::Display for Token {
@@ -27,7 +27,12 @@ impl fmt::Display for Token {
 }
 
 impl Token {
-    pub fn new(token_type: TokenType, lexeme: String, literal: Option<Literal>, line: u64) -> Self {
+    pub fn new(
+        token_type: TokenType,
+        lexeme: String,
+        literal: Option<Literal>,
+        line: usize,
+    ) -> Self {
         Token {
             token_type,
             lexeme,
